server/requirements.txt
notebooklm-py[browser]
